updates structure for facet history functions/calls
