+++ lib/fetchContent.js
@@ -2,6 +2,8 @@
 
 const fetch = require('node-fetch');
 const debug = require('debug')('lib:fetchContent');
+const parameters = require('../helpers/parameters');
+const time = require('../helpers/time');
 const CAPI_KEY = process.env.CAPI_KEY;
 if (!CAPI_KEY) {
 	throw new Error('ERROR: CAPI_KEY not specified in env');
@@ -301,9 +303,138 @@ function getArticle(uuid) {
 		});
 }
 
+async function searchFacetHistory(params){
+	const MAX_FACETS			= 100;
+	const MAX_INTERVAL			= 5;
+	const MAX_INTERVAL_NUM		= 5;
+	const DEFAULT_PERIOD		= 'days';
+	const DEFAULT_FACETS		= 10;
+	const DEFAULT_INTERVAL		= 1;
+	const DEFAULT_INTERVAL_NUM	= 5;
+
+	try {
+		const searchFacet		= params.facet;
+		const searchPeriod		= parameters.check('string', params.period, 0, 0, ['minutes','hours','days'], 'days');
+		const intervaledFacets 	= [];
+		const recentTopFacets	= [];
+		const facetHistory		= {};
+
+		let numInterval 		= parameters.check('int', params.interval, 1, MAX_INTERVAL, null, DEFAULT_INTERVAL);
+		let numIntervals 		= parameters.check('int', params.numInterval, 1, MAX_INTERVAL_NUM, null, DEFAULT_INTERVAL_NUM);
+		let numFacetItems		= parameters.check('int', params.maxFacets, 1, MAX_FACETS, null, DEFAULT_FACETS);
+		let fullDateTime		= time.getDatetimeRange(searchPeriod, (numInterval * numIntervals), 0);
+		let resultFacets		= {
+			description : "Returns metrics for facet numbers over the time period specificed in the params of the query",
+			requestParams : {
+				facet : searchFacet,
+				period : searchPeriod,
+				interval : numInterval,
+				numInterval : numIntervals,
+				maxFacets : numFacetItems,
+			},
+			datetimeRange : {
+				start : fullDateTime.first,
+				end : fullDateTime.next,
+			}
+		};
+
+
+		const searches = createDateTimeRangeQueryStrings(searchPeriod, numInterval, numIntervals, searchFacet);
+		const queryResults = await searchSequence(searches);
+
+
+		// get facet elements for all returned days
+		queryResults.forEach(result => {
+			let facet = result.sapiObj.results[0].facets;
+			let indexCount = result.sapiObj.results[0].indexCount;
+
+			if(indexCount > 0 && facet !== undefined){
+				facet.forEach( facetElement => {
+					if(facetElement.name === searchFacet){
+						intervaledFacets.push( facetElement.facetElements );
+					}
+				});
+			}
+		});
+
+
+		// stop if no results found
+		if(intervaledFacets.length === 0){
+			res.json({
+				error: "No results found for this set of queries"
+			});
+			return;
+		}
+
+		// get most recent segement of top facets
+		if(numFacetItems >= intervaledFacets[0].length){
+			numFacetItems = intervaledFacets[0].length;
+		}
+
+		for (var i = 0; i < numFacetItems; i++) {
+			let facet = intervaledFacets[0][i];
+
+			if(facet.hasOwnProperty('name')){
+				recentTopFacets.push(facet.name);
+			}
+		}
+
+
+		// find the numbers for each of today's top facet elements
+		recentTopFacets.forEach( topFacet => {
+			facetHistory[topFacet] = [];
+
+			intervaledFacets.forEach( day => {
+				const facetValue = day.filter(item => item.name == topFacet);
+
+				if(facetValue[0] !== undefined && facetValue[0].hasOwnProperty('count')){
+					facetHistory[topFacet].push(facetValue[0].count);
+				} else {
+					facetHistory[topFacet].push(0);
+				}
+			});
+		});
+
+
+		resultFacets[searchFacet] = facetHistory;
+		return resultFacets;
+
+	} catch (err) {
+		console.log('err: ' + err);
+	}
+}
+
+function createDateTimeRangeQueryStrings(period, numInterval, numIntervals, searchFacet){
+	let queries = [];
+	for (let i = 0; i <= numIntervals; i++) {
+		queries.push( createFacetQueryString(period, numInterval, searchFacet, i) );
+	}
+	return queries; 
+}
+
+function createFacetQueryString(period, invterval, facetName, offset){
+
+	if( facetName === "topics" || facetName === "people" || facetName === "organisations" ){
+		const datetimeRange = time.getDatetimeRange(period, invterval, offset);
+
+		return {
+			"queryString": `lastPublishDateTime:>${datetimeRange.next} AND lastPublishDateTime:<${datetimeRange.first}`,
+			"maxResults" : 1,
+			"facets" : {
+				"names" : [facetName, facetName + "Id"],
+				"maxElements" : -1
+			}
+		};	
+	}
+
+	console.log("createFacetQueryString: Incorrect facetName passed");
+	return "";
+}
+
 module.exports = {
 	search,
 	searchDeeper,
 	searchSequence,
-	getArticle
+	getArticle,
+	searchFacetHistory
 };
