+++ modules/facet.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const fetchContent = require('../lib/fetchContent');
 const directly = require('../helpers/directly');
 const CAPI_CONCURRENCE = process.env.hasOwnProperty('CAPI_CONCURRENCE')
@@ -8,7 +10,11 @@ function searchBySequence(searches){
 	return fetchContent.searchSequence(searches);
 }
 
+function searchForFacetHistory(params){
+	return fetchContent.searchFacetHistory(params);
+}
 
 module.exports = {
-	searchBySequence
+	searchBySequence,
+	searchForFacetHistory
 };
